+++ sample/components/tabs.js
@@ -1,5 +1,20 @@
 import React, { Component } from 'react';
+import PropTypes from 'prop-types';
 import { Spacing, Tabs, Text } from 'react-elemental';
+
+const CustomTabOption = ({ children }) => (
+  <Spacing size="tiny" top bottom padding>
+    <Spacing size="small" left right padding>
+      <Text size="kilo" color="gray60" uppercase bold>
+        {children}
+      </Text>
+    </Spacing>
+  </Spacing>
+);
+
+CustomTabOption.propTypes = {
+  children: PropTypes.string.isRequired,
+};
 
 export default class SampleTabs extends Component {
   state = {};
@@ -7,7 +22,7 @@ export default class SampleTabs extends Component {
   handleChange = (key) => (selected) => this.setState({ [key]: selected });
 
   render() {
-    const { simple, complex = 'options' } = this.state;
+    const { simple, complex = 'options', custom = 'can', invert = 'highlight' } = this.state;
 
     return (
       <div>
@@ -23,19 +38,26 @@ export default class SampleTabs extends Component {
         <Spacing size="huge" bottom>
           <Spacing bottom>
             <Text size="iota" color="gray70" uppercase bold>
-              Generic
+              Primary
             </Text>
           </Spacing>
 
+          <Tabs
+            options={[
+              { value: 'one', label: 'one' },
+              { value: 'two', label: 'two' },
+              { value: 'three', label: 'three' },
+            ]}
+            value={simple}
+            onChange={this.handleChange('simple')}
+          />
+        </Spacing>
+
+        <Spacing size="huge" bottom>
           <Spacing bottom>
-            <Tabs
-              options={[
-                { value: 'one', label: 'one' },
-                { value: 'two', label: 'two' },
-              ]}
-              value={simple}
-              onChange={this.handleChange('simple')}
-            />
+            <Text size="iota" color="gray70" uppercase bold>
+              Secondary
+            </Text>
           </Spacing>
 
           <Spacing bottom>
@@ -51,8 +73,45 @@ export default class SampleTabs extends Component {
               ]}
               value={complex}
               onChange={this.handleChange('complex')}
+              secondary
             />
           </Spacing>
+
+          <Spacing size="small" bottom>
+            <Text>
+              The tab options can also be condensed to fit their contents exactly, rather than
+              distributing the width evenly across all options.
+            </Text>
+          </Spacing>
+
+          <Spacing bottom>
+            <Tabs
+              options={[
+                { value: 'labels', label: <CustomTabOption>Labels</CustomTabOption> },
+                { value: 'can', label: <CustomTabOption>can</CustomTabOption> },
+                { value: 'be', label: <CustomTabOption>be</CustomTabOption> },
+                { value: 'arbitrary', label: <CustomTabOption>arbitrary</CustomTabOption> },
+                { value: 'children', label: <CustomTabOption>children</CustomTabOption> },
+              ]}
+              value={custom}
+              onChange={this.handleChange('custom')}
+              secondary
+              fit
+            />
+          </Spacing>
+
+          <Tabs
+            options={[
+              { value: 'highlight', label: <CustomTabOption>Highlight</CustomTabOption> },
+              { value: 'selector', label: <CustomTabOption>Selector</CustomTabOption> },
+              { value: 'inverted', label: <CustomTabOption>inverted</CustomTabOption> },
+            ]}
+            value={invert}
+            onChange={this.handleChange('invert')}
+            secondary
+            fit
+            invert
+          />
         </Spacing>
       </div>
     );